Update: Unit tests for f_utf and relating changes or bug fixes.
Fix several problems exposed by unit tests.
Fix several unit tests to work as expected due to problems with the data files.

At some point I seem to have diverged from always ensuring that the f_utf_char_t is always big-endian.
I probably got lost in ensuring the differences between big and little endian that I ended up making the f_utf_char_t act little-endian in cases where host is little-endian.
The f_utf_char_t must always be big-endian.
However, there are cases where the big and little endian behavior must be processed.
Break up the macros into having "_be" and "_le" to make this possible.

The iscntrl() check return value needs to be explicitly handled to ensure that only F_false or F_true is returned.
This is already fixed in one function already.
Apply the existing fix to the other function.

The f_utf_char_t should be seen as a single character rather than a stream of bytes.
Unit tests now treat any non-zero value after the designated width as invalid.
The is valid checking code now tests for this invalid case.

The unit tests are improved.
Test for F_true and F_false rather than calling assert_true() and assert_false().
Error bits and other status codes were previously passing when they should fail due to the use of assert_true() and assert_false().

This commit changes the byte order of the f_utf_char_t.
This will break code such as the code used in the utf8 program.
A follow up commit is necessary to fix any byte order problems.

+++ level_0/f_utf/tests/unit/c/data-utf.h
@@ -311,13 +311,13 @@ extern FILE *data__bytesequence_file_open__zero_width(void);
  * @return
  *   positive number on success where number represents bytes read.
  *   0 on success and end of file is reached.
- *   0 on failure.
+ *   -1 on failure.
  *
  * @see atoll()
  * @see getline()
  * @see htonl()
  */
-extern uint8_t data__bytesequence_get_line(FILE * const file, f_utf_char_t * const character);
+extern ssize_t data__bytesequence_get_line(FILE * const file, f_utf_char_t * const character);
 
 #ifdef __cplusplus
 } // extern "C"
